Sandbox Background Complete
I had to manually set groups of tiles to get them all in the right place, to the best-looking hundredth of a unit.  Looks decent.  We'll just leave the sprites as they are in the background and build on top of that.
TODO: Make corner path sprite (if we want to), change camera size and shift to center.